Added the files
The Daily_data.csv is an example of what an API call will look like.
THe Test.ipynb is the code that makes the prediction, before making it an .py I’d like to confront on the main functionalities and how they interact with our interface for the users. I think that everything is settuped in the right way but I’ll leave it open to changes.
The Valley_Boundaries.csv is the map that enables us to select the model to pick when making the predictions based on the location that the user wants to choose.